[MaL] Move all Kelvin vector code in single place.
