+++ ProcessLib/Deformation/BMatrixPolicy.h
@@ -13,46 +13,6 @@
 
 namespace ProcessLib
 {
-/// Kelvin vector dimensions for given displacement dimension.
-template <int DisplacementDim>
-struct KelvinVectorDimensions;
-
-template <>
-struct KelvinVectorDimensions<2>
-{
-    static int const value = 4;
-};
-
-template <>
-struct KelvinVectorDimensions<3>
-{
-    static int const value = 6;
-};
-
-//
-// Kelvin vector and matrix templates for given displacement dimension.
-//
-
-/// \todo Maybe better to move the KelvinVector/MatrixType into
-/// MaterialLib/SolidModels/KelvinVector.h.
-
-/// Kelvin vector type for given displacement dimension.
-/// \note The Eigen vector is always a fixed size vector in contrast to the
-/// BMatrixPolicyType::KelvinVectorType.
-template <int DisplacementDim>
-using KelvinVectorType =
-    Eigen::Matrix<double, KelvinVectorDimensions<DisplacementDim>::value, 1,
-                  Eigen::ColMajor>;
-
-/// Kelvin matrix type for given displacement dimension.
-/// \note The Eigen matrix is always a fixed size matrix in contrast to the
-/// BMatrixPolicyType::KelvinMatrixType.
-template <int DisplacementDim>
-using KelvinMatrixType =
-    Eigen::Matrix<double, KelvinVectorDimensions<DisplacementDim>::value,
-                  KelvinVectorDimensions<DisplacementDim>::value,
-                  Eigen::RowMajor>;
-
 /// An implementation of B-Matrix policy using same matrix and vector types
 /// (fixed size or dynamic) as in the ShapeMatrixPolicyType.
 template <typename ShapeFunction, unsigned DisplacementDim>
